+++ tests/test_agatesql.py
@@ -104,6 +104,22 @@ class TestSQL(agate.AgateTestCase):
         for dialect in ['mysql', 'postgresql', 'sqlite']:
             self.table.to_sql_create_statement('test_table', dialect=dialect)
 
+    def test_to_sql_create_statement_zero_width(self):
+        rows = (
+            (1, ''),
+            (2, ''),
+        )
+        column_names = ['id', 'name']
+        column_types = [agate.Number(), agate.Text()]
+
+        table = agate.Table(rows, column_names, column_types)
+
+        statement = table.to_sql_create_statement('test_table', db_schema='test_schema', dialect='mysql')
+
+        self.assertIn('CREATE TABLE test_schema.test_table', statement)
+        self.assertIn('id DECIMAL(38, 0) NOT NULL,', statement)
+        self.assertIn('name VARCHAR(1)', statement)
+
     def test_sql_query_simple(self):
         results = self.table.sql_query('select * from agate')